Remove Go checker from normal builds
