+++ internal/stdlib/rebind_test.go
@@ -3,9 +3,6 @@ package stdlib
 import (
 	"testing"
 
-	"github.com/osty/osty/internal/check"
-	"github.com/osty/osty/internal/diag"
-	"github.com/osty/osty/internal/parser"
 	"github.com/osty/osty/internal/resolve"
 )
 
@@ -51,76 +48,6 @@ func TestPreludeBuiltinRebindings(t *testing.T) {
 			if sym.Kind != resolve.SymVariant {
 				t.Fatalf("std.%s.%s kind = %s, want variant", tc.module, name, sym.Kind)
 			}
-		}
-	}
-}
-
-func TestQualifiedStdlibTypesSharePreludeIdentity(t *testing.T) {
-	src := []byte(`use std.collections
-use std.option
-use std.result
-use std.error
-
-pub fn collectionsRoundTrip(xs: List<Int>, m: Map<String, Int>, s: Set<Int>) -> Int {
-    let qxs: collections.List<Int> = xs
-    let bxs: List<Int> = qxs
-    let qm: collections.Map<String, Int> = m
-    let bm: Map<String, Int> = qm
-    let qs: collections.Set<Int> = s
-    let bs: Set<Int> = qs
-    bxs.len() + bm.len() + bs.len()
-}
-
-pub fn optionRoundTrip(x: Int?) -> Int? {
-    let q: option.Option<Int> = x
-    let b: Int? = q
-    b
-}
-
-pub fn resultRoundTrip(r: Result<Int, Error>) -> result.Result<Int, error.Error> {
-    let q: result.Result<Int, error.Error> = r
-    let b: Result<Int, Error> = q
-    b
-}
-
-pub fn errorRoundTrip(e: Error) -> error.Error {
-    let q: error.Error = e
-    let b: Error = q
-    b
-}
-`)
-	file, parseDiags := parser.ParseDiagnostics(src)
-	for _, d := range parseDiags {
-		if d.Severity == diag.Error {
-			t.Fatalf("parse: %s", d.Error())
-		}
-	}
-	reg := Load()
-	res := resolve.FileWithStdlib(file, resolve.NewPrelude(), reg)
-	chk := check.File(file, res, check.Opts{
-		Primitives:    reg.Primitives,
-		ResultMethods: reg.ResultMethods,
-	})
-	for _, d := range append(res.Diags, chk.Diags...) {
-		if d.Severity == diag.Error {
-			t.Fatalf("unexpected diagnostic: %s", d.Error())
-		}
-	}
-}
-
-func TestCollectionsPureBodiesTypeCheckAfterRebind(t *testing.T) {
-	reg := Load()
-	mod := reg.Modules["collections"]
-	if mod == nil || mod.Package == nil {
-		t.Fatal("std.collections not loaded")
-	}
-	chk := check.Package(mod.Package, &resolve.PackageResult{PackageScope: mod.Package.PkgScope}, check.Opts{
-		Primitives:    reg.Primitives,
-		ResultMethods: reg.ResultMethods,
-	})
-	for _, d := range chk.Diags {
-		if d.Severity == diag.Error {
-			t.Fatalf("std.collections should type-check: %s", d.Error())
 		}
 	}
 }
